Guarantee that wr is not lost for a spurious temporary merge
If a queue Q1 is merged with an already-shared queue Q2, then Q1
may easily be freed soon after this event. If this happens and
the merge lasts little, then Q1 may be created again at a time
instant at which it had still to enjoy weight raising. This
commit checks that, and restores weight raising for the queue
if it is the case.

Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Signed-off-by: Albert I <krascgq@outlook.co.id>
Signed-off-by: Saurabh Charde <saurabhchardereal@gmail.com>
Signed-off-by: Ruega <purelbatako@gmail.com>